src/requirements.txt
azure-core==1.26.1
azure-storage-blob==12.14.1
black==22.12.0
certifi==2022.12.7
cffi==1.14.6
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
Flask==2.2.2
idna==3.4
importlib-metadata==5.2.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msrest==0.7.1
mypy-extensions==0.4.3
oauthlib==3.2.2
pathspec==0.10.3
platformdirs==2.6.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
